updating worker connectors
