--- workers/unity/Assets/Playground/Scripts/UI/ConnectionScreenController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ce6e8e48beb442439555421a825c197
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
