--- workers/unity/Assets/Playground/Scripts/Worker/IMobileConnectionController.cs
@@ -1,13 +1,0 @@
-using Improbable.Gdk.Core;
-
-namespace Playground
-{
-    public interface IMobileConnectionController
-    {
-        string IpAddress { get; set; }
-
-        ConnectionScreenController ConnectionScreenController { get; set; }
-
-        void TryConnectAsync(ConnectionService connectionService);
-    }
-}
