--- workers/unity/Assets/Playground/Scripts/Worker/IMobileConnectionController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: aeee29f4d51e42d2bc960d47ace1150e
-timeCreated: 1539705431
