--- workers/unity/Assets/Playground/Scripts/Worker/iOSClientWorkerConnector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b30b8628bcc74fb3ac4de050f53566a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
